removes overlapping files, only keeping essentials
