--- Sources/Extensions/UIKit/UICollectionViewExtensions.swift
@@ -1,17 +1,0 @@
-//
-//  Copyright © FINN.no AS, Inc. All rights reserved.
-//
-
-import UIKit
-
-// extension UICollectionViewCell: Identifiable {}
-//
-// extension UICollectionView {
-//    func register(_ cellClass: UICollectionViewCell.Type) {
-//        register(cellClass.self, forCellWithReuseIdentifier: cellClass.reuseIdentifier)
-//    }
-//
-//    func dequeue<T>(_ cellClass: T.Type, for indexPath: IndexPath) -> T where T: UICollectionViewCell {
-//        return dequeueReusableCell(withReuseIdentifier: T.reuseIdentifier, for: indexPath) as! T
-//    }
-// }
