--- Sources/Extensions/UIKit/UITableViewExtensions.swift
@@ -1,17 +1,0 @@
-//
-//  Copyright © FINN.no AS, Inc. All rights reserved.
-//
-
-import UIKit
-
-// extension UITableViewCell: Identifiable {}
-//
-// extension UITableView {
-//    func register(_ cellClass: UITableViewCell.Type) {
-//        register(cellClass.self, forCellReuseIdentifier: cellClass.reuseIdentifier)
-//    }
-//
-//    func dequeue<T>(_ cellClass: T.Type, for indexPath: IndexPath) -> T where T: UITableViewCell {
-//        return dequeueReusableCell(withIdentifier: cellClass.reuseIdentifier, for: indexPath) as! T
-//    }
-// }
