--- Sources/Finnivers/DNA/Spacing/Spacing.swift
@@ -1,25 +1,0 @@
-//
-//  Copyright © FINN.no AS, Inc. All rights reserved.
-//
-
-import CoreGraphics
-
-extension CGFloat {
-    /// Separation of 2 points.
-    static let verySmallSpacing: CGFloat = 2
-
-    /// Separation of 4 points.
-    static let smallSpacing: CGFloat = 4
-
-    /// Separation of 8 points.
-    static let mediumSpacing: CGFloat = 8
-
-    /// Separation of 16 points.
-    static let mediumLargeSpacing: CGFloat = 16
-
-    /// Separation of 32 points.
-    static let largeSpacing: CGFloat = 32
-
-    /// Separation of 64 points.
-    static let veryLargeSpacing: CGFloat = 64
-}
